requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo[srv]==4.7.1
motor==3.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
Pillow==10.3.0
aiofiles==23.2.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
certifi==2024.2.2
openai